Simplify the RecommendedFeeds with breakpoint components
